Se hicieron refactorizaciones necesarias para un desarrollo escalable
