+++ config/dbEventoRepositorio.js
@@ -1,10 +1,10 @@
 const Pool = require("pg").Pool;
 
 const pool = new Pool({
-  user: "hsazteibnsnquc",
-  password: "96c44f19b6a31a67521c2fa65c9233544ed1d7d5388367c6d9ff4c22c940a340", //use your pass my friend
-  database: "d5mjf648gc2p7f",
-  host: "ec2-54-156-24-159.compute-1.amazonaws.com",
+  user: "hgpmlfhmjxvnfr",
+  password: "e3fcf341e4ff4a68075b951e1c9a75239afaa42d7eccc3e9c7db81bda6c77a05", //use your pass my friend
+  database: "d966qfatdj765h",
+  host: "ec2-54-173-138-144.compute-1.amazonaws.com",
   port: 5432,
   ssl: {
     rejectUnauthorized: false,
@@ -22,14 +22,31 @@ class DbEventoRepositorio {
     return eventos;
   }
 
-  async create_evento(data)
-  {
-      const {nombre_evento,descripcion_evento,modalidad_evento,lugar_evento,fecha_evento,proyecto,estado}= data
-      const new_evento = await pool.query("INSERT INTO public.eventos(nombre_evento,descripcion_evento,modalidad_evento,lugar_evento,fecha_evento,proyecto,estado) VALUES ($1, $2, $3, $4, $5, $6, $7)",
-      [nombre_evento,descripcion_evento,modalidad_evento,lugar_evento,fecha_evento,proyecto,estado]);
-      return new_evento
+  async create_evento(data) {
+    const {
+      nombre_evento,
+      descripcion_evento,
+      modalidad_evento,
+      lugar_evento,
+      fecha_evento,
+      proyecto,
+      estado,
+    } = data;
+    const new_evento = await pool.query(
+      "INSERT INTO public.eventos(nombre_evento,descripcion_evento,modalidad_evento,lugar_evento,fecha_evento,proyecto,estado) VALUES ($1, $2, $3, $4, $5, $6, $7)",
+      [
+        nombre_evento,
+        descripcion_evento,
+        modalidad_evento,
+        lugar_evento,
+        fecha_evento,
+        proyecto,
+        estado,
+      ]
+    );
+    return new_evento;
   }
-  
+
   async get_evento(data) {
     const { id } = data.params;
     const evento = await pool.query(
@@ -38,23 +55,31 @@ class DbEventoRepositorio {
     );
     return evento;
   }
-  
+
   async delete_evento(id) {
-    console.log(id)
-    const eliminar_evento = await pool.query('DELETE FROM public.eventos WHERE id = $1', 
-    [id]);
+    console.log(id);
+    const eliminar_evento = await pool.query(
+      "DELETE FROM public.eventos WHERE id = $1",
+      [id]
+    );
     return eliminar_evento;
   }
   //Archivar evento
-  async update_evento_estado1(data){
+  async update_evento_estado1(data) {
     console.log(data.id);
-      const actualizar_estado = await pool.query('UPDATE public.eventos SET estado = 0  WHERE id = $1', [data]);   
+    const actualizar_estado = await pool.query(
+      "UPDATE public.eventos SET estado = 0  WHERE id = $1",
+      [data]
+    );
     return actualizar_estado;
   }
   //Mostrar Evento
-  async update_evento_estado2(data){
+  async update_evento_estado2(data) {
     console.log(data.id);
-      const actualizar_estado = await pool.query('UPDATE public.eventos SET estado = 1  WHERE id = $1', [data]);   
+    const actualizar_estado = await pool.query(
+      "UPDATE public.eventos SET estado = 1  WHERE id = $1",
+      [data]
+    );
     return actualizar_estado;
   }
